refactor: separate imports for client/server code in game package
