+++ apps/game/client/index.ts
@@ -3,3 +3,4 @@ export * from "./game-state-client";
 export * from "./build-version-context";
 export * from "./area/area-debug-ui";
 export * from "./area/area-scene";
+export * from "./useRPC";
