--- apps/game/index.ts
@@ -1,17 +1,0 @@
-export type * from "./server/game-state";
-export type * from "./server/character/schema";
-export type * from "./server/npc/schema";
-export type * from "./server/traits/actor";
-export type * from "./server/traits/appearance";
-export type * from "./server/traits/combat";
-export type * from "./server/traits/movement";
-export type * from "./server/area/schema";
-export * from "./server/";
-export * from "./client/";
-export * from "./shared/area/area-resource";
-export * from "./shared/area/tiled-resource";
-export * from "./shared/area/area-id";
-export * from "./shared/client-view-distance-rect";
-export * from "./rpc";
-export * from "./user/index";
-export * from "./rpc.slice";
