+++ apps/game/server/index.ts
@@ -18,3 +18,10 @@ export * from "./traits/movement";
 export * from "./traits/appearance";
 export * from "./traits/actor";
 export * from "./client-registry";
+export * from "../shared/client-view-distance-rect";
+export * from "../shared/area/area-id";
+export * from "../shared/area/area-resource";
+export * from "../shared/area/tiled-resource";
+export * from "../user/index";
+export * from "./rpc";
+export * from "./rpc.slice";
