+++ apps/game/server/rpc.slice.ts
@@ -1,0 +1,7 @@
+import { areaRouterSlice, npcRouterSlice, characterRouterSlice } from ".";
+
+export const gameRpcSlice = {
+  ...characterRouterSlice,
+  ...areaRouterSlice,
+  ...npcRouterSlice,
+};
