+++ apps/server/src/etc/rate-limiter-middleware.ts
@@ -1,5 +1,5 @@
 import { RPCError } from "@mp/rpc";
-import { ctxSessionId, rpc } from "@mp/game";
+import { ctxSessionId, rpc } from "@mp/game/server";
 import { RateLimiter } from "@mp/rate-limiter";
 
 const globalRequestLimit = new RateLimiter({ points: 20, duration: 1 });
