+++ apps/server/src/metrics/user.ts
@@ -2,7 +2,7 @@ import type { MetricsRegistry } from "@mp/telemetry/prom";
 import { MetricsGague } from "@mp/telemetry/prom";
 import { recordValues } from "@mp/std";
 import type { PatchStateMachine } from "@mp/sync";
-import type { ClientRegistry, GameState } from "@mp/game";
+import type { ClientRegistry, GameState } from "@mp/game/server";
 
 export function collectUserMetrics(
   registry: MetricsRegistry,
